+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - ActiveRecord has now been demoted to being a data provider, it is no longer tightly coupled with Repository, this means __we could have multiple data providers__!
   - Repository is now much more robust
   - As a result of robustness, it fixed an issue where updating an existing entity with new data records from mapped attributes will not persist the new data records correctly
+  - `#save` and `#destroy` now support accepting an array as their argument
 
 ## 0.10.1 [2013-03-16]